backend/requirements.txt
fastapi[standard]        
psycopg2-binary   
passlib[bcrypt]
pydantic-settings
uvicorn
